Deployment/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-resumable-media==2.5.0
google-crc32c==1.5.0
grpcio==1.59.0
grpcio-status==1.59.0
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9
six==1.16.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorboard==2.13.0
tensorflow-hub==0.14.0
tensorflow-text==2.13.0
absl-py==1.4.0
astunparse==1.6.3
flatbuffers==23.5.26
gast==0.4.0
google-pasta==0.2.0
h5py==3.9.0
keras==2.13.1
libclang==16.0.6
numpy==1.24.3
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
scipy==1.11.2
termcolor==2.3.0
typing_extensions==4.7.1
wrapt==1.15.0
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
gunicorn==21.2.0
python-dotenv==1.0.0
gdown==4.7.1 